hybrid mamba implementation

+++ scripts/train_causal.py
@@ -294,7 +294,12 @@ def training_state(model_path, model, rank):
 
 
 def main():
-    torch.set_default_dtype(torch.bfloat16)
+    torch.set_default_dtype(torch.float32) # mamba
+    #torch.set_default_dtype(torch.bfloat16) # llama
+
+    # import causal_conv1d_cuda
+    # print("causal_conv1d_cuda loaded successfully")
+
 
     print0("Loading model...")
     model = models.get_model(
